# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AdEarn API
  description: |
    Intent-matched cashback advertising platform.
    Users declare purchase intent, see matched ads, complete purchases,
    and receive cashback split across 4 pools: liquid, savings, parent fund, charity.

    ## Authentication
    Most routes require a Bearer JWT (RS256). Obtain tokens via `POST /auth/verify-otp`.
    The `role` claim in the JWT controls access:
    - `consumer` — Feed, Attribution, Wallet, Reviews
    - `advertiser` — Advertiser routes
    - `admin` — Admin routes

    ## Response Envelope
    All responses follow:
    ```json
    { "success": true, "data": <T> }        // success
    { "success": false, "error": { "code": "...", "message": "..." } }  // error
    ```

    ## Financial Invariants
    1. Cashback distribution is always one atomic PostgreSQL transaction.
    2. Advertiser billing happens only after cashback is committed.
    3. Every webhook is idempotent — double-processing is architecturally impossible.
  version: 1.0.0
  contact:
    name: Abhishek K
    email: admin@adearn.in
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://adearn-api.railway.app/api/v1
    description: Production (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT with `role` claim (`consumer` | `advertiser` | `admin`)
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Authentication required

    # ─── Domain schemas ───────────────────────────────────────────────────────
...
    FraudQueueRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_mobile:
          type: string
          example: "9876543210"
        campaign_name:
          type: string
          example: Mamaearth Vitamin C Serum
        purchase_amount:
          type: string
          description: Purchase amount in INR (decimal string)
          example: "1499.00"
        cashback_amount:
          type: string
          description: Cashback amount in INR (decimal string)
          example: "44.97"
        fraud_score:
          type: string
          description: Fraud score 0.0–1.0 (decimal string)
          example: "0.85"
        created_at:
          type: string
          format: date-time
    AdminUserRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobile:
          type: string
          example: "9876543210"
        name:
          type: string
          example: Priya Sharma
        role:
          type: string
          enum: [consumer, advertiser, admin]
          example: consumer
        kyc_status:
          type: string
          enum: [pending, verified, rejected]
          example: verified
        is_active:
          type: boolean
          example: true
        fraud_flags:
          type: integer
          description: Number of fraud flags raised on this account
          example: 0
        created_at:
          type: string
          format: date-time
    PendingAdvertiserRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
          example: Mamaearth Pvt Ltd
        gst_number:
          type: string
          example: 27AADCB2230M1ZT
        contact_email:
          type: string
          format: email
          example: ads@mamaearth.in
        contact_mobile:
          type: string
          example: "9123456789"
        pledge_signed:
          type: boolean
          example: true
        status:
          type: string
          enum: [pending, approved, rejected]
          example: pending
        created_at:
          type: string
          format: date-time
    FinancialSummary:
      type: object
      properties:
        total_cashback_paid:
          type: string
          description: Total cashback paid out in INR (decimal string)
          example: "128450.00"
        total_under_review:
          type: string
          description: Cashback currently under fraud review (decimal string)
          example: "3200.00"
        total_liquid:
          type: string
          description: Aggregate liquid pool balance across all users (decimal string)
          example: "51380.00"
        total_savings:
          type: string
          description: Aggregate savings pool balance (decimal string)
          example: "38535.00"
        total_parent_pending:
          type: string
          description: Aggregate parent fund balance awaiting disbursement (decimal string)
          example: "25690.00"
        total_charity_pending:
          type: string
          description: Aggregate charity pool balance awaiting disbursement (decimal string)
          example: "12845.00"
        active_users:
          type: integer
          description: Number of users with at least one conversion
          example: 342
    AuditLogRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
          nullable: true
        actor_mobile:
          type: string
          nullable: true
          example: "9000000000"
        action:
          type: string
          example: FRAUD_APPROVE
        entity_type:
          type: string
          example: cashback_transaction
        entity_id:
          type: string
          nullable: true
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        before_state:
          type: object
          nullable: true
          additionalProperties: true
        after_state:
          type: object
          nullable: true
          additionalProperties: true
        ip_address:
          type: string
          nullable: true
          example: 203.0.113.42
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Insufficient role — valid JWT but wrong role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Admin role required
    UnprocessableEntity:
      description: Validation error — request body or query params invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "liquid_pct + savings_pct + parent_pct + charity_pct must equal 100"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found

# ─── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /auth/request-otp:
    post:
      tags: [Auth]
      summary: Request OTP
      description: |
        Sends a 6-digit OTP to the supplied mobile number.
        In development (`OTP_MOCK=true`) the OTP is always `123456` — no SMS sent.
        Rate limited: 3 requests per 10 minutes per mobile.
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile]
              properties:
                mobile:
                  type: string
                  pattern: '^[6-9]\d{9}$'
                  description: 10-digit Indian mobile number
                  example: "9876543210"
      responses:
        '200':
          description: OTP dispatched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
              example:
                success: true
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Too many OTP requests. Try again in 10 minutes.
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP and obtain tokens
      description: |
        Verifies the OTP and returns JWT access + refresh tokens.
        Creates a new user account on first login.
        Rate limited: 5 attempts per 5 minutes per mobile.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, otp]
              properties:
                mobile:
                  type: string
                  pattern: '^[6-9]\d{9}$'
                  example: "9876543210"
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "123456"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        description: RS256 JWT — expires in 24h
                        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                      refresh_token:
                        type: string
                        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
                            nullable: true
                            example: Priya Sharma
                          role:
                            type: string
                            enum: [consumer, advertiser, admin]
                            example: consumer
        '401':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: INVALID_OTP
                  message: OTP is incorrect or has expired
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Profile ──────────────────────────────────────────────────────────────
...
  /admin/fraud-queue:
    get:
      tags: [Admin]
      summary: List fraud review queue
      description: Returns cashback transactions flagged for fraud review (score > 0.8 threshold).
      operationId: getFraudQueue
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Fraud queue entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FraudQueueRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/fraud-queue/{id}/approve:
    put:
      tags: [Admin]
      summary: Approve or reject a fraud queue entry
      description: |
        Approves (`approved: true`) or rejects (`approved: false`) a flagged cashback.
        Approval releases the cashback to the consumer's pools.
        Rejection reverses the cashback and credits the advertiser.
        All actions written to the audit log.
      operationId: resolveFraudQueueEntry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Fraud queue entry ID (cashback transaction ID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Fraud queue entry resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      approved:
                        type: boolean
                        example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      description: Returns paginated user list for admin management. Supports role and status filters.
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [consumer, advertiser, admin]
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{id}/suspend:
    put:
      tags: [Admin]
      summary: Suspend or reinstate a user
      description: |
        Sets `is_active = false` (suspend) or `true` (reinstate).
        Suspended users cannot log in or trigger new cashback.
        Action written to audit log.
      operationId: suspendUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspended]
              properties:
                suspended:
                  type: boolean
                  description: true = suspend, false = reinstate
                  example: true
      responses:
        '200':
          description: User suspension status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      suspended:
                        type: boolean
                        example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/advertisers:
    get:
      tags: [Admin]
      summary: List pending advertiser applications
      description: Returns advertiser accounts awaiting admin approval.
      operationId: adminListAdvertisers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
            default: pending
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Advertiser application list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PendingAdvertiserRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/advertisers/{id}/approve:
    put:
      tags: [Admin]
      summary: Approve or reject advertiser application
      description: |
        Approves (`approved: true`) or rejects (`approved: false`) an advertiser application.
        Approved advertisers can create campaigns immediately.
        Action written to audit log.
      operationId: approveAdvertiser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Advertiser ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Advertiser application resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      approved:
                        type: boolean
                        example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/financials:
    get:
      tags: [Admin]
      summary: Get platform financial summary
      description: Returns aggregate financial metrics across all pools, users and cashback status categories.
      operationId: getFinancials
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Financial summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/FinancialSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/audit-log:
    get:
      tags: [Admin]
      summary: Get audit log
      description: |
        Returns a filterable audit trail of all admin actions (fraud approvals,
        user suspensions, advertiser approvals, etc.).
      operationId: getAuditLog
      security:
        - bearerAuth: []
      parameters:
        - name: action
          in: query
          description: Filter by action type (e.g. FRAUD_APPROVE, USER_SUSPEND)
          schema:
            type: string
            example: FRAUD_APPROVE
        - name: entity_type
          in: query
          description: Filter by entity type (e.g. cashback_transaction, user)
          schema:
            type: string
            example: cashback_transaction
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...